requirements.txt
pandas==2.2.1
requests==2.31.0
geopandas==0.14.3
python-dotenv==1.0.1
